Remove links in title when showing them in the backlinks buffer
Previously if a file has a link in its title, eg.

    Unit Testing ([[notes:elisp.org][Emacs Lisp]])

Then inside elisp.org, we'll see a backlink from that file that looks
like:

    [[notes:unit-testing-elisp.org][Unit Testing ([[notes:elisp.org][Emacs Lisp]])]]

which is invalid Org syntax. (You cannot nest links inside links.)

This function keeps only the description (or the bare link) around, so
that we get this instead:

    [[notes:unit-testing-elisp.org][Unit Testing (Emacs Lisp)]]