Don't format field values in APA style
I want names to be displayed as-is. If we're writing a research paper
it might be appropriate to use APA as that is an accepted standard,
but in my personal notes I prefer to not have mangled names.

(I merged commit 5236917 instead of my no-apa branch of
org-roam-bibtex. This reapplies the changes in that branch.)