remove the "link-type" argument from org-roam-format-link
That argument was meant to communicate whether the link should be
absolute or relative, I think. When I changed it to abbreviate links
based on org-link-abbrev-alist it's already unused.

This fixes an issue where file renames fail to trigger link
replacement. I thought an unused input argument (one prefixed with an
underline) can still be passed in...?