fix: actually fix Markdown information extraction this time
Because extraction functions rely on `major-mode` to determine which
variant to use, we need to set the right major mode.

The logic surrounding auto-mode-alist in set-auto-mode isn't available
as a separate function (and set-auto-mode doesn't allow passing in a
FILE argument), so I had to copy the core logic here.