minaduki-lit: fix default type being "???" instead of the file name
See org-refresh-category-properties: the default category is the base
name of the buffer file name. If buffer-file-name is nil, it uses
"???" instead.

To make it grab the right default category, we have to set
buffer-file-name; we also have to reset it so that kill-buffer doesn't
start prompting whether we should kill a temporary buffer in
with-temp-buffer.