add a new command that lists all broken links, replacing org-roam-doctor
While org-roam-doctor also checks for syntax, I don't see a need for
it. My user base is realistically just myself, or maybe if I refine
this enough, some people looking for an alternative to org-roam; but I
don't have the ability to polish this enough to the point where I can
accommodate beginners like org-roam does.