test: really "fix" the mysterious failing spec
I observed that if I switch to using org-roam--list-files-elisp, I can
reproduce the failure on my system. Ripgrep is installed on my system
but not in CI, making it a possible reason why the behaviors could be
different.

Then after enabling progress messages, I realized that for some
inexplicable reason, the tests would run while the cache is still
building.

I do not understand how this can happen. Buttercup's before-all code
seems entirely synchronous. From bin/buttercup to
buttercup-run-discover to buttercup-run, nothing spawns a thread, a
process, or does any trickery with run-at-time. The before-all block
is wrapped in a function (without being immediately invoked) when the
specs are first defined, so they shouldn't be run at that time either.

But, err, we can just run it again to ensure it has actually built the
cache, in a section where we're 100% sure it will run in order.

If this still fails on CI I might just go crazy.