new method of determining if a reference is unlinked
Previously references were considered unlinked when their context do
not contain the path to the current file. This breaks for relative
links.

Now a reference is considered unlinked if it comes from a file that
does not have a link to the current file. This relies on a file-based
workflow, but otherwise remains fast and accurate enough.