Fix modification detection for some files with editorconfig-mode
Given:

- a file encoded with LF line endings
- in a repository where the .editorconfig calls for CRLF line ends
- (example of a repository like this:
   https://github.com/obsidianmd/obsidian-docs/ )

then compute-content-hash would return different values within the
buffer versus when passed with the file.

So, uh, fix that.

(Original confusion below)
Some files, in my case half of Obsidian's Help vault, are always
treated as new.

I initially thought this is because they use CRLF, but this is
sometimes not reproducable.

Then I thought it's because they do not have a final newline, but this
also sometimes doesn't reproduce.

Then copying it elsewhere, with the file contents being identical (I
checked with sha256sum), somehow made it go away as well.

It's... maddening, but since there is this possibility and removing it
does not hurt (and actually avoids some computation), I'm still going
to commit this.