Speed up caching for YAML front matter
yaml-parse-string can be slow, and we were doing it repeatedly. This
speeds it up by caching the entire front matter parse result when
generating the DB.

Before this, building the DB from scratch took over a second for a
Markdown file with YAML front matter. After, it's finally measured in
files per second again, and not seconds per file.