fix(info): work around unreliable dir entry
The file name in dir entry can actually deviate from the file as a
mistake. For example, libmicrohttpd's tutorial's dir entry starts with:

   * libmicrohttpdtutorial: (libmicrohttpd). […]

I think this is a bug on their end and have reported it as
https://bugs.gnunet.org/view.php?id=7928, but this is something that I
still have to work around. The "File: " line is probably good enough;
if it turns out to not be the case, we can just use the file path.