vectors
json-serialize does not allow arrays to be passed in as lists. As in,
if you try to serialize a list (that isn't an alist or a plist) to
JSON, it will error out.

I think it's still worth it to convert lists to vectors when
applicable and use json-serialize instead of switching to json-encode.