Remove track-files-outside-vaults
It was made to

- make it possible to see backlinks to files outside vaults from inside
  a vault
- make it possible to link to files outside a vault when inside a vault

but it also has the side effect of making it too easy to create links
from outside a vault into a vault, which is unlikely to be valid.

Before I (find the time to) figure out a way to achieve the first two
without also making editing Org and JSON files miserable, this should be
removed for now.

(This is what I mean when I say I expect myself to be the only user.)