fix: don't fail even if the comp-setup variables are not defined
(defvar <variable>) along doesn't make that variable not error when
referred to, it just silences the bytecomp warning. We have to actually
initialize it to nil to make it not error.

A better approach might be to not emit the setup if the variable isn't
defined at macro expansion time, or to do a runtime check.

We can probably also just remove the defvar since it'd just become an
unused lexical variable in an environment where the completion system
being configured isn't present, which is harmless.