Fix being hard to access new notes in early hours of a day
...when org-extend-today-until is having an effect.

With the variable at 2, then 1 AM would be considered the previous day,
but then if I create a new note `minaduki/open-diary-entry` would not
consider that note.

Ideally this matching should also take org-extend-today-until into
account, so that even if a note is created at eg. 003000 (and o-e-t-u is
2) that would still be considered as from the previous day. But due to
the way I match dates from the cache database that would require more
effort. Having both *not* honor it is still better than just one part
honoring it.