fix(lit)!!!!: broken bibliography key insertion
it was being inserted **and then immediately cleared by
minaduki-db::insert-note-lit-entries** because we assumed that function
can safely clear the file's entries and then re-add it if it's present.
For blibliography files, that is not the case.

With that erroneous clearing gone, when bibliography entries are removed
they would then linger. Passing update-p = true to ::insert-lit-entries
helps clear existing entries at the actually appropriate time.