fix: initial info for a new note of a citekey sometimes not being written
If I create a new note for a citekey, then decided to delete the file
for some reason, the file has a stale entry in the database and the next
time I create a new note for the same citekey we would end up doing
find-file on an empty file. (In Doom Emacs, this is extra weird because
the file-templates system then kicks in.)

Instead, we should only use find-file if the file actually exists on
disk.