Switch to Eask
Cask is already not working on my device, so this makes sense.

The setup experience is *horrible*, the effects of which on me can be
seen in what I wrote in the Eask file.

But otherwise the commands make sense. Zero extra setup buttercup is
nice. Builtin formatting is nice.